Remove guess counter reset when switching modes and wrap scroll in a setTimeout to make it async
Without the async, the page would begin scrolling to the Y coordinate of the complete message. After the contents of the table loads, the complete message's Y coordinate is updated, but the page is still mid-scroll to its old Y coordinate. The async scrolls the page properly to where the complete message is.